refactor: split TextBubble into TextWrapper and BubbleDrawer submodules
Split text_bubble.rb (379 lines) into three files under 250 lines each:
- text_bubble.rb (152 lines): core draw/position logic
- text_bubble/text_wrapper.rb (76 lines): CJK-aware text wrapping
- text_bubble/bubble_drawer.rb (146 lines): bubble shape drawing

Refs #426
