+++ ruby/smalruby3/lib/smalruby3/render/text_bubble.rb
@@ -2,6 +2,8 @@
 
 require "sdl2"
 require_relative "font_support"
+require_relative "text_bubble/text_wrapper"
+require_relative "text_bubble/bubble_drawer"
 
 module Smalruby3
   module Render
@@ -9,6 +11,8 @@ module Smalruby3
     # Draws directly via SDL2 renderer (not surface-based).
     class TextBubble
       include FontSupport
+      include TextWrapper
+      include BubbleDrawer
 
       MAX_TEXT_LENGTH = 330
       MAX_LINE_WIDTH = 170
@@ -70,237 +74,6 @@ module Smalruby3
         SDL2::TTF.init
         path = find_font
         @font = SDL2::TTF.open(path, FONT_SIZE) if path
-      end
-
-      # --- Text wrapping (character-by-character for CJK support) ---
-
-      def wrap_text(text)
-        return [""] if text.empty?
-
-        lines = []
-        text.split("\n").each do |para|
-          if para.empty?
-            lines << ""
-          else
-            wrap_paragraph(para, lines)
-          end
-        end
-        lines.empty? ? [""] : lines
-      end
-
-      def wrap_paragraph(para, lines)
-        current = ""
-        buf = "" # word buffer for non-CJK
-
-        para.each_char do |ch|
-          if ch == " "
-            # Flush buffer + space
-            buf += ch
-          elsif cjk?(ch)
-            # Flush any buffered word first
-            current = flush_word(buf, current, lines) unless buf.empty?
-            buf = ""
-            # Try adding CJK char
-            test = current + ch
-            w, = @font.size_text(test)
-            if w <= MAX_LINE_WIDTH
-              current = test
-            else
-              lines << current unless current.empty?
-              current = ch
-            end
-          else
-            buf += ch
-          end
-        end
-
-        # Flush remaining buffer
-        current = flush_word(buf, current, lines) unless buf.empty?
-        lines << current unless current.empty?
-      end
-
-      def flush_word(word, current, lines)
-        test = current + word
-        w, = @font.size_text(test)
-        if w <= MAX_LINE_WIDTH
-          test
-        else
-          lines << current unless current.empty?
-          word.lstrip
-        end
-      end
-
-      def cjk?(ch)
-        cp = ch.ord
-        cp.between?(0x3000, 0x9FFF) ||
-          cp.between?(0xF900, 0xFAFF) ||
-          cp.between?(0xFF00, 0xFFEF)
-      end
-
-      # --- Bubble drawing ---
-
-      def draw_bubble(x, y, w, h, type, on_right)
-        @sdl_renderer.draw_blend_mode = SDL2::BlendMode::BLEND
-
-        # Stroke first (behind fill, matching Scratch's stroke→fill order)
-        @sdl_renderer.draw_color = STROKE_COLOR
-        draw_rounded_rect(x, y, w, h, CORNER_RADIUS)
-
-        # Fill over stroke
-        inset = STROKE_WIDTH / 2
-        @sdl_renderer.draw_color = FILL_COLOR
-        draw_rounded_rect(x + inset, y + inset,
-          w - STROKE_WIDTH, h - STROKE_WIDTH,
-          [CORNER_RADIUS - inset, 1].max)
-
-        # Tail below bubble
-        draw_tail(x, y, w, h, type, on_right)
-
-        # Erase the border between bubble and tail by overdrawing
-        # the junction area with fill color
-        erase_tail_junction(x, y, w, h, on_right, type)
-      end
-
-      def draw_rounded_rect(x, y, w, h, r)
-        r = [r, w / 2, h / 2].min
-        @sdl_renderer.fill_rect(SDL2::Rect.new(x + r, y, w - 2 * r, h))
-        @sdl_renderer.fill_rect(SDL2::Rect.new(x, y + r, r, h - 2 * r))
-        @sdl_renderer.fill_rect(SDL2::Rect.new(x + w - r, y + r, r, h - 2 * r))
-        fill_circle(x + r, y + r, r)
-        fill_circle(x + w - r - 1, y + r, r)
-        fill_circle(x + r, y + h - r - 1, r)
-        fill_circle(x + w - r - 1, y + h - r - 1, r)
-      end
-
-      def erase_tail_junction(x, y, w, h, on_right, type)
-        return if type == :think # think has detached circles, no junction
-
-        # The say tail top edge (y=0) spans from x=-16 to x=0 relative to tail_x.
-        # With dir multiplier, calculate the screen-space range of the junction.
-        jw = 16
-        jx = if on_right
-          # dir=-1: tail_x to tail_x+16
-          x + CORNER_RADIUS
-        else
-          # dir=1: tail_x-16 to tail_x
-          x + w - CORNER_RADIUS - 16
-        end
-        @sdl_renderer.draw_color = FILL_COLOR
-        # Erase the border at the junction (bottom of bubble where tail attaches)
-        @sdl_renderer.fill_rect(SDL2::Rect.new(jx, y + h - STROKE_WIDTH,
-          jw, STROKE_WIDTH + 1))
-      end
-
-      def draw_tail(x, y, w, h, type, on_right)
-        # Scratch draws tail at (paddedWidth - CORNER_RADIUS, paddedHeight)
-        # = near the RIGHT-BOTTOM corner of the bubble.
-        # When pointsLeft (=on_right), the entire canvas is flipped horizontally,
-        # so the tail appears at the LEFT-BOTTOM corner.
-        #
-        # We simulate the flip by choosing tail_x:
-        # on_right (pointsLeft): tail at LEFT side → origin near left-bottom
-        # !on_right: tail at RIGHT side → origin near right-bottom
-        if on_right
-          # Tail origin at left side, tail extends further left toward sprite
-          tail_x = x + CORNER_RADIUS
-          tail_y = y + h - 1
-          draw_tail_shape(tail_x, tail_y, -1, type)
-        else
-          # Tail origin at right side, tail extends further right toward sprite
-          tail_x = x + w - CORNER_RADIUS
-          tail_y = y + h - 1
-          draw_tail_shape(tail_x, tail_y, 1, type)
-        end
-      end
-
-      def draw_tail_shape(tx, ty, dir, type)
-        if type == :say
-          draw_say_tail(tx, ty, dir)
-        else
-          draw_think_tail(tx, ty, dir)
-        end
-      end
-
-      def draw_say_tail(tx, ty, dir)
-        # Scratch tail bezier path (origin = bubble's right-bottom corner area):
-        #   Start at (0, 0)
-        #   Bezier1: cp1=(0,4) cp2=(4,8) end=(4,10) → curves right then down
-        #   Arc tip: (4,10)→(4,12)→(2,12) r=2 → rounded tip
-        #   Bezier2: cp1=(-1,12) cp2=(-11,8) end=(-16,0) → curves back left-up
-        #
-        # We sample the two bezier curves to get left/right edges at each Y.
-        # Right edge = bezier1, Left edge = bezier2 (reversed)
-
-        # Pre-compute edge X at each scanline by sampling the beziers
-        right_edge = Array.new(TAIL_HEIGHT + 1, 0)
-        left_edge = Array.new(TAIL_HEIGHT + 1, 999)
-
-        # Sample bezier1: (0,0)→(0,4)→(4,8)→(4,10) for right edge
-        # Sample bezier2: (-16,0)→(-11,8)→(-1,12)→(2,12) for left edge
-        steps = TAIL_HEIGHT * 8
-        (0..steps).each do |i|
-          t = i.to_f / steps
-          # Bezier1 right edge
-          rx = cubic_bezier(t, 0, 0, 4, 4)
-          ry = cubic_bezier(t, 0, 4, 8, 10)
-          yi = ry.round.clamp(0, TAIL_HEIGHT)
-          right_edge[yi] = [right_edge[yi], rx.round].max
-
-          # Bezier2 left edge (from (-16,0) to tip area)
-          lx = cubic_bezier(t, -16, -11, -1, 2)
-          ly = cubic_bezier(t, 0, 8, 12, 12)
-          yi = ly.round.clamp(0, TAIL_HEIGHT)
-          left_edge[yi] = [left_edge[yi], lx.round].min
-        end
-
-        # Clamp unvisited rows
-        (0..TAIL_HEIGHT).each do |dy|
-          left_edge[dy] = [left_edge[dy], right_edge[dy]].min
-        end
-
-        # Stroke first (behind fill, matching Scratch's stroke→fill order)
-        # Draw stroke with expanded edges so it shows around the fill
-        @sdl_renderer.draw_color = STROKE_COLOR
-        sw = STROKE_WIDTH / 2
-        (0..TAIL_HEIGHT).each do |dy|
-          x1 = tx + dir * (left_edge[dy] - sw)
-          x2 = tx + dir * (right_edge[dy] + sw)
-          x1, x2 = x2, x1 if x1 > x2
-          @sdl_renderer.draw_line(x1, ty + dy, x2, ty + dy) if x2 >= x1
-        end
-
-        # Fill over stroke
-        @sdl_renderer.draw_color = FILL_COLOR
-        (0..TAIL_HEIGHT).each do |dy|
-          x1 = tx + dir * left_edge[dy]
-          x2 = tx + dir * right_edge[dy]
-          x1, x2 = x2, x1 if x1 > x2
-          @sdl_renderer.draw_line(x1, ty + dy, x2, ty + dy) if x2 >= x1
-        end
-      end
-
-      def cubic_bezier(t, p0, p1, p2, p3)
-        mt = 1 - t
-        mt * mt * mt * p0 + 3 * mt * mt * t * p1 + 3 * mt * t * t * p2 + t * t * t * p3
-      end
-
-      def draw_think_tail(tx, ty, dir)
-        # Scratch's think tail (coords relative to bubble bottom-right corner):
-        #   Half-circle at (-16, 0) r=4 — attached to bubble bottom
-        #   Circle at (-9.25, 7.25) r=2.25 — middle
-        #   Circle at (-1.5, 9.5) r=1.5 — smallest, furthest down
-        # Note: positive Y = downward (toward sprite)
-        bubbles = [
-          [tx + dir * -16, ty + 1, 4],    # Largest, attached to bubble
-          [tx + dir * -9, ty + 7, 2],     # Medium
-          [tx + dir * -2, ty + 10, 1]     # Smallest, furthest from bubble
-        ]
-        bubbles.each do |cx, cy, r|
-          @sdl_renderer.draw_color = STROKE_COLOR
-          fill_circle(cx, cy, r + 1)
-          @sdl_renderer.draw_color = FILL_COLOR
-          fill_circle(cx, cy, r)
-        end
       end
 
       # --- Text rendering ---
